move useOverlayContainer to hooks dir
