+++ lib/Modal/Modal.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 import { Transition, TransitionGroup } from 'react-transition-group';
 import { uniqueId, noop } from 'lodash';
 import { getFirstFocusable } from '../../util/getFocusableElements';
-import useOverlayContainer from '../../util/useOverlayContainer';
+import useOverlayContainer from '../../hooks/useOverlayContainer';
 import WrappingElement from './WrappingElement';
 import { OVERLAY_CONTAINER_ID } from '../../util/consts';
 import IconButton from '../IconButton';
