+++ lib/Popper/Popper.js
@@ -1,10 +1,10 @@
-import React, { useContext, forwardRef, useRef } from 'react';
+import React, { useContext, forwardRef } from 'react';
 import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import PopperJS from 'popper.js';
 import css from './Popper.css';
 import { OVERLAY_CONTAINER_ID } from '../../util/consts';
-import useOverlayContainer from '../../util/useOverlayContainer';
+import useOverlayContainer from '../../hooks/useOverlayContainer';
 import StripesOverlayContext from '../../util/StripesOverlayContext';
 
 export const OVERLAY_MODIFIERS = {
